Finish initial version of State and move to main package
